fix: TournamentTeam.TotalPoints int → double para compatibilidad con nueva formula
CS0266: no se puede asignar double a int al recalcular leaderboard.
Migración actualizada para alterar también TournamentTeams.TotalPoints.

+++ src/WarzoneTournament.Domain/Entities/TournamentTeam.cs
@@ -10,7 +10,7 @@ public class TournamentTeam : BaseEntity
     public bool CheckedIn { get; set; } = false;
     public DateTime? CheckInTime { get; set; }
     public bool IsEliminated { get; set; } = false;
-    public int TotalPoints { get; set; } = 0;
+    public double TotalPoints { get; set; } = 0;
     public int TotalKills { get; set; } = 0;
     public int CurrentRank { get; set; } = 0;
